[EC-58] chore: utils 파일 간소화
